refactor(wage): fix typo on file related to wage "garnishment"
